check too long msg

+++ code/src/Application/UseCase/Bot/BotUseCase.php
@@ -152,6 +152,9 @@ final class BotUseCase
                 if ($inlineKeyboardData === 'settings_menu') {
                     $allTelegramMessageByUser = $this->telegramMessageRepository->getAllMessageByChatId($telegramUser->telegram_chat_id);
                     foreach ($allTelegramMessageByUser as $msg) {
+                        if($msg == end($allTelegramMessageByUser)){
+                           continue;
+                        }
                         $this->telegram::deleteMessage($msg->chat_id, $msg->message_id);
                         $this->telegramMessageRepository->deleteByMessageId($msg->message_id);
                     }
